Restore 0.24 console map scrolling behavior
In 2fe5a39f5e4 code that was only used to initialize the map view, prior
to the commit, was converted to set the region and cursor position on
every render. The result was a change in behavior, where previously
scrolling only moved the viewport when it hit the edges of the map,
after this commit it effectively scrolled the viewpoint on any movement
on big enough maps, keeping the cursor in the center. It's possible this
was even intentional, not sure, because the new behavior resembled what
happens in local tiles and webtiles in the map view (this would at least
make it easy to miss relative to non-console map view). However, this
behavior is not a good fit for the console map view; The map changing on
any movement of the cursor is (IMO) much harder for the user to process
than just changing the cursor position, and the required full-screen
redraw every step is slower over ssh. (To be clear, I actually think
this first point applies to the other two UIs as well, but one step at a
time.)

This commit restores the 0.24 behavior with (what I think is) a small
improvement: scrolling within a few squares of the edge rather than at
the edge will trigger movement of the viewport on the map, always
keeping a 3 tile margin. (0.24 and earlier scrolled at the edge.)