Move some text for some artefacts to art-data.txt.
Create DESCRIP and DBRAND fields in art-data.txt. These fields are used
in the description of an artefact. DBRAND extends the brand description
and should be used for special brands. DESCRIP is used for sundry
properties.

The staff of Olgreb, Storm Bow, Maxwell's Thermic Engine, salamander
hide armour, macabre finger necklace, and the glaive of the Guard are
all adjusted to use these new data fields.

I've put this in the description below the "excrutiating wounds"
description, but this doesn't matter as the spell cannot be used on
artefacts.

I haven't moved the Warlock's Mirror description, as this field only
works if the text to insert is fixed.

Add + as a "start of line" character to show that the contents of a line
in art-data.txt continues the preceding (string) field. Where, with
whitespace, a space is added between the old line and the new one, with
+, a newline is added.

Add a description of the fields to the top of art-data.txt.

[ Committer's note: Squashed and rebased, revised the commit message,
  tweaked some documentation and moved formatting newlines to
  describe.cc only (no extra newlines from art-data.pl). Closes #1717 ]