Move some text for some artefact weapons from describe.cc to art-data.txt.
Create a DESCRIP field in art-data.txt. This is inserted into the help text for an artefact weapon. Use it for staff of Olgreb, Storm Bow and Maxwell's Thermic Engine.

I've put this in the description below the "excrutiating wounds" description, but this doesn't matter as the spell cannot be used on artefacts.

I haven't moved the Warlock's Mirror description, as this field only works if the text to insert is fixed.