Refactor targeting to minimize console redraws
38af612a8a cleaned up the direction chooser code-wise quite a bit, but
it had unintended side-effects for console. In particular, after this
commit, the entire window was fully cleared and redrawn on every
keypress. This isn't very noticeable on local builds, but it is
extremely heavy over ssh and can (especially with lower ping) cause bad
flicker when targeting while playing online. (It also is going to have a
non-trivial impact, in aggregate, on ttyrec size.) This commit refactors
the widget wrapping so that most of the time a full screen redraw is not
needed; only the parts of the view that need to be redrawn get redrawn,
the message window is only redrawn if there are messages, and the
sidebar is only redrawn when a full screen redraw is triggered (e.g.
initially, or by opening and closing the help window). This code may or
may not be a bit hacky relative to more normal Widget classes.

This should involve no change for local tiles. However, as far as I can
tell the new OverlayWidget class works fine in local tiles, if doing all
this redrawing were to be a problem there. Should also not impact
webtiles targeting.

(cherry picked from commit 0dd03a7ff89a8cfd5b48dfdc0a004616898175c7)