Revert "Increase broad axe delay"
This reverts commit 81720017eba9a398c36b19f0d9c4cbe01de8cb61.

In the commit message for 81720017eb, its pointed out that broad axes
are too cheap relative to battle axes for their rarity. Part of the
issue on later analysis of skill points required, is that *shields* were
too cheap to add on in a penalty-free way. Commit 0b64aa26 increased the
shield penalty. To swing a 20 skill broad axe at min-delay with a kite
shield now takes 21,000 skill points on a human; compared to the 13,950
of a battle axe. For an 18 skill broad axe the cost is 17,850; still
more than the battle axe (since the trade is for -2 base damage in
exchange for the defenses of a shield, it should cost more), but not
1.5 times more expensive.

To repeat some of the discussion from 0b64aa26, prior to that commit the
cost to swing an 18 skill borad axe at min-delay was 12,200, cheaper
than the broad axe. Addressing the disparity via broad axe delay was one
solution, but made broad axes a non-use nobrainer for small species (the
same base damage is available 1-handed at 16 skill from a morning star
for them!); broad axes are still pretty expensive for a small species, but
should be more attractive at 18 delay when the shield changes are also
considered.