Don't let tabbing throw nets with autofight_throw/nomove
I'm not sure it's the best path to go down to keep adding small
hard-coded exceptions like this, but this change brings nets in line
with other quiver actions affected by 006ef2cfcc4; they won't be
triggered by autofight commands that may fire the quiver. The biggest
case where this matters is on a Gl start with one of the relevant
options set. In a very small poll of players it seems like most people
have been using autoinscriptions for this and don't want these commands
to ever auto-shoot nets anyways. As with that previous commit,
CMD_AUTOFIRE itself will still work.

It's less clear what to do with other ammo starts where players may not
want to tab-fire, e.g. poison darts, dispersal boomerangs.
Autoinscribing `!Q` or `!f` is currently an option but that's very
fiddly and more for advanced rc file users.