Rework monster placement in a minivault
The vault called simply "pond" places over D:10- and all of Lair, but
has pretty unambitious monster placement. Except for the fact that its
insubstantial wisps got replaced with will-o-the-wisps!  Aside from this
unfortunate wisp storm, the monster list didn't scale well with
placement. This commit reworks the placement, conditioning it on depth
and refining the two Lair and Swamp themes it has.

Lair and Swamp monster sets are used for D:10-12 and Lair, but only the
Swamp set is used for D:13-, since the amphibious Lair monsters aren't
an interesting enough set by that point. The Lair set starts from D:10
mostly easier Lair spawns, but ramps this up through Lair and throws
torpor snails into the mix by Lair:4. The Swamp set starts as just
vampire mosquitoes with light swamp drakeage on D:10, but as depth
increases adds swamp dragons, hydra, and eventually will-o-the-wisps.