Remove *Curse
Random curses on equip existed on exactly three unrands (and no randarts).
It was on Scythe of Curses purely for theme, on Obsidian Axe to make it
not a complete no-brainer to swap off the weapon when mesmerized, and on
the Necklace of Bloodlust to discourage swaps when you'd rather not be
randomly berserked.

*Curse wasn't an extremely great mechanism for limiting the latter two.
The randomness of how many scrolls of remove curse were generated meant
that the restriction was sometimes tight and sometimes irrelevant, and
there wasn't any real tactical element to the cost.

So, let's remove it. Amulets are now slow to swap, so the Necklace of
Bloodlust no longer has a *Drain downside at all. The Scythe of Curses
is now {drain, *Drain}, which seems very thematic if not mechanically
necessary, and the obsidian axe likewise gets *Drain. (It could be fun
to have a 'Compel' or 'Sticky' property that makes these weapons slow
to swap, compelling the user like the One Ring... but not right now.)
This should hopefully provide a more interesting set of tactical choices.

Curses!