Create a dbrand field in struct unrandart_entry. This is treated in the same way as the descrip one, except that a blank line is always left after each line in it.
In art-data.pl, DBRAND is treated in the same way as DESCRIP, except that any newlines within a string are doubled, giving the required blank line after each one.

Add + as a "start of line" character to show that the contents of a line in art-data.txt continues the preceding (string) field. Where, with whitespace, a space is added between the old line and the new one, with +, a newline is added.

Add a description of the fields to the top of art-data.txt.

Change the artefact definitions in art-data.txt to take account of these changes. There is no longer a need to write \n in the file.