Remove Airstrike's bonus damage against fliers
Against monsters, it was just another type of elemental vulnerability, but
against players, it was almost never tactically interesting and heavily
penalised forgetting to toggle permanent sources of flight off. Rather than
make it behave inconsistently against monsters, remove the bonus entirely -
the increased damage against targets in open space is a more interesting
bonus to focus the spell on (although the way it works should probably be
adjusted/simplified).