Quiet down storm clouds
Storm clouds are loud (lightning bolt loud!) because Grunt made them
that way in February 2014 and there wasn't much other commentary about
it at the time.

They're the only cloud that makes noise, and appear in five places:
- Around Bai in second form in her cloud ring
- Random panlord cloud rings
- Robe of Clouds clouds
- Condenser vane clouds
- Qazlal clouds

For monsters, r-i indicated that the loudness of Bai was mildly desired
but not a primary part of the design; with randlords the cloud rings are
very much a "don't melee me" effect and not for noise.

For the two evoked player sources of clouds, the condenser vane has
skill training and xp gating to balance it. The Robe of Clouds probably
needs another re-work, it does rely on storm cloud noise for balance
somewhat.

Finally, Qazlal stormclouds have made Q much louder than designed, even
very early on, making the noise conduct much harsher than it should be.
Q's rumble is min ( piety / 10, 16 ) noise, lightning bolt is 25. 1* Q
followers weren't supposed to be so loud.

Making all clouds quiet will make it easier to attach appropriate noise
(or other cost or drawback) to their source according to mechanical
niche, instead of this one cloud type being SO LOUD.