Skip elven twin stuff on dismissal/reset
On a delver start, if the los zap wipes out one elven twin, the other is
supposed to go too. But, the code in elven_twin_died still ran on the
first to go and caused the second one to be woken and berserked etc;
depending on the timing this could lead to messaging as in mantis 12480.
I'm pretty sure that there's no natural way for these kill types to
apply, so this should affect only the delver start case.