Skip some unique fineff messaging on turn 0 (12346)
This is for the case where a delver lands in los of a pikel/kirke, where
it doesn't make sense to do the messaging related to freeing the
follower band. I haven't tested this case very directly because the seed
in mantis 12346 doesn't place pikel any more (and I haven't tried to go
back to a version where it does), so it's possible there's more tweaking
needed. In principle, the bands should get zapped too (like dowan and
duvessa) but I don't *think* this is happening right now.