Remove some water from mumra_spider_spiderweb
The vault has several spiders trapped behind deep water, as they can
no longer cling to walls.

This commit frees the spiders and reduces the number of water and deep
water tiles in the vault. Also, it removes an escape hatch in the
inner area, since the vault is not a teleport closet anymore.