Reduce the experience cost of high-level skills
The cost of increasing a skill from 20 to 27 was more than the cost of increasing a skill from 0 to 20, making
training extremely high level skills (e.g. for level 9 spells or slow two-handed weapons) a poor investment for most
of the game. This commit makes the implicit quadratic-with-break-point formula explicit, and divides the break point
contributions to skill costs above level 9 by 2. This has the following implications:
* Skill costs up to level 9 are unaffected;
* The new cost of skill level 15 is half-way between the old costs of level 14 and 15;
* The new cost of skill level 20 is approximately the same as the old cost of skill level 19;
* The new cost of skill level 27 is approximately the same as the old cost of skill level 25;
* The decreases in costs are roughly equivalent to a +1 increase in aptitude by skill level 23,
  and the impact is always smaller than a +2 increase in aptitude.

[Committer's note: code refactored.]

Closes #1733.