Add more per-species titles
Seems fun! All of these are intended to be not too annoying to get
for folks who like to collect titles, but they're not in general
designed as something to steer folks towards a race's strengths -
just fun little acknowledgements of something the player's done.

I desperately wanted the formicid title to be 'transdimensional tunneler',
but it was far too long...