Explicitly show ally foes
This shows ally targeting info: (i) on the ally, (ii) on the target
monsters, and (iii) as a mi flag on the short description (without
specifics). For simple cases this won't show anything surprising at all
to experienced players, but I think it still is useful info to new
players. For complex cases, it may be useful to see e.g. that an ally is
pathing through a monster that's not what you'd hope it to be targeting.

I suspect this hasn't been done before because of information overload
issues. And monster descriptions definitely have been accruing
information overload for the last few versions! However, I think I have
converted to team info, and I do think this specific bit of info is
useful for many cases. It also may have not been done before because
there are cases where what is shown is weird and confusing. However,
this reflects weird and confusing ally *behaviour*, which is still
frustrating to players when it is detected. This commit may make it
easier to detect, which is perhaps the path towards less confusing ally
behavior. I won't be surprised if it turns up some bugs too...

It doesn't seem impossible to me that this could turn out to be a bad
idea to show after all, but we'll see.