Distinguish between "1 poisoned" and "all poisoned" in the monster list.
Previously, if there were 6 sheep, the monster list would show:
(poisoned) if 1 had been poisoned;
(2 poisoned) if 2 had; and
(poisoned) if all had.

This changes the first case to (1 poisoned). It displays (poisoned), as now, if only one sheep is left.

This makes it easier to see when (in this case) a sheep needs more poison.