Rein in some autoswitch logic
In particular, don't apply autoswitch logic to ammo gain/loss unless it
is an entire stack of ammo. This partially addresses 12488: what is
happening there is that because firing changes stack size, the
autoswitch code was getting called on any firing event. It's not a
complete fix because `F` should still never affect the fire history for
autoswitch purposes.