Partial refactor of throw_item_no_quiver (12488)
The end goal is to have this better integrated with quiver code, but
this commit only moves a bit towards this. The main target of this
commit, convulated though it may seem, is to prevent `F` from updating
the global quiver history and avoid the issue in mantis 12488
altogether. This commit does it by using a custom action_cycler for
throwing via `F`, which indirectly keeps the history from being updated
and minorly improves the interface for `F` targeting (and then most of
the commit is restructuring so that this can be done more easily).