Limit stair exclusions to stairs
Altars and shops also have stair directions, but stair exclusion
semantics doesn't make any sense for them (and regular exclusions should
still work if you want that). Also, because this is called a lot in
rendering, checking the stair direction of altars was a weird little
hotspot because it has to iterate through possible altar features (2% of
weight during an early game qw run with a profiler attached, which is comparable to all of console rendering).

Is there a use case for stair exclusion semantics acting on
transporters? I'm not sure if that even worked but this probably
prevents it if it did. If people complain I'll add them back.