Memoize some terrain functions
This fixes up a bunch of existing small hotspots by using the built in
memoize code; somewhat redundant with the previous commit, which dealt
with a stair direction call in a different way already. Really, the
consequential change here is the feat_is_altar change, but I've also
memoized a few key things that were downstream from this just because
it's easy to do. The suppress blood check is heavily called on a
webtiles build, so seems to merit converting to a more standard terrain
function and memoizing as well (though really it was a hotspot because
of the altar check).