Simplify spectral weapons
Simplify the effect. The weapon pops out when you attack, immediately
follows up, and then sticks around until the end of your next turn.
If you hit again, everything repeats. Otherwise, it vanishes.

The design here is of an ego with great damage but that has some
positioning constraints, and that still has the potential drawback of
damage sharing. (If the positioning constraints turn out to be enough to
make it interesting, we can junk the damage sharing.)

I've just removed monster spectral ego for now, but it'd be nice to have
it back eventually.