Reforge the Chains V: Bind skill knowledge to curses
This commit reimagines Ashenzari curses to offer specific knowledge (in
exchange for binding the slot). Currently Ash picks two skills and the
player can either take the curse and the knowledge, or ignore it. Ash
will eventually offer a new curse.

This choice can hit the player in a variety of ways: are the skills
offered too good to pass up, thus ending up on an otherwise not great
item for a long time? Are the skills not ones the player anticipates
using but they're desparate for piety so they'll put a not great curse
on a great item? Both are good, and occasionally the player will delight
in getting the opportunity to put a good curse on a good item they want
to keep!

With the variety of curses, it remains possible for an Ash character to
be a jack of all trades, but also to become highly focused in something
that might otherwise be hard for their species! Exciting possibilities
abound, that will need some further number tuning.