Turn wear-id into floor-id
Following much the same reasoning as in f6d4cf81.

Especially now that curses are gone (RIP), there's very little incentive not
to wear-ID items. (This was mostly true before too, but it's extra obvious
now.) Instead, just identify equippable items when you step on them. This does
remove certain categories of surprise from randarts (it's *Contam!), but
players will live.

No comp for this in terms of ?id drops, since that item was better spent on
consumables for the period of the game in which ?id is in finite supply.
(So no difference.)

Closes #1770.