Create a new vision_iterator (PleasingFungus)
A handful of places in the code performed a radius_iterator iteration
over LOS_NONE and then immediately filtered for you.see_cell to respect
scrying and wizard xray vision effects. This iterator encapsulates the
idiom of iterating over an actor's vision (permitting actor::see_cell to
override the considerations of LOS checks).