Activate Ash skill boost from 0* (mdonais)
The skill boost requires curses to take advantage of; at 0* and only one
curse the skill point multiplier is 4. Compared to 24 for one curse at
6* this isn't that noticable.

In exchange, the play experience is clearer: all of the messages about
curses being tied to skills don't have to have caveats, and a Monk
actually gets something immeidate out of using their first curse.