Consolidate and clean up some body part name code
This tries to reduce a bunch of code duplication while making some body
part name cases more consistent, and also reduce reliance on
species-specific checks.

Biggest changes to blade hands code, which is/was fairly involved, but
hopefully all changes are improvements. The commit removes the general
blade hands name ("scythe-like blade") that was used in various cases,
awkwardly, and uses a clean version of the regular hand name. For ice
form, we use "Ice paws" (instead of "Ice fists") in the UI -- the form
data already used "paws" as the hand name (for e.g. ring wearing), so
this improves consistency. For statue form, I used the regular hand name
instead of "fists"; maybe this should still be customized for the "hand"
case (but it reduces duplication of similar code).

This stuff was all complex enough that there may well be corner cases
that this commit breaks...