Reduce explicit species checks: tengu
This commit is part of a series that attempts to more cleanly treat
species as bundles of traits/behaviors that have their own well-defined
semantics, and minimize the use of explicit species checks in the code.

Tengu are easy. A previous commit cleaned up some superficial code, and
this one uses the tengu flight mutation for a bunch of cases rather than
checking the species directly.