Vaults monsters reworks!
Unify ironheart and ironbrands into ironbounds, per request. Hopefully this
won't cause too much confusion.

Ironbound Frostheart:
Spellcasting monsters with a new Creeping Frost spell, which deals cold
damage to and flash-freezes all enemies adjacent to walls. They also have
an AF_COLD touch attack. Overall, they're intended to be a mid-tier threat
which plays with the terrain of Vaults (where there tend to be a lot of
walls around!) and which fills a gap in rC-checking monsters there.

Ironbound Thunderhulks:
Playing around with new targeting constraints, this imposing Vaults
foe tosses smite-targeted lightning at its foes. Just one catch -
it can't hit the player if they get within 1-2 tiles of it! Seems
like a fun way to play with the mobs and spaces of Vaults.

Ironwhip Beastmaster
Instead of having beasts spawn in the vaults on their own, let's add
on a Vaults thematic strength: bigger fights with synergizing bands. The
Ironwhip Beastmasters are the ones who have taken the late, rare Lair
monsters to the Vaults.

Beastmasters come with a "Goad Beasts" natural ability, which gives all
beasts in LOS a bonus to energy equal to one EUT_MOVE. This is less
pronounced than haste other or a torpor snail aura for basic combat;
however they come with bands of trampling dire elephants or fire
breathing lindwurms.

Here the goad helps the monsters in the band manouver around one another
to better get to the player. For trampling monsters this does have a
different mechanical impact than a torpor snail aura: after a successful
trample that moves the player the remaining goaded elephants have a
better chance at getting a hit.

Beastmasters also spawn with nets, because what's a good beast master
without a net.

Move ugly things to wizard bands in Vaults

This is more for Lore than for mechanics. Having plain ugly things
appear either with wizards in small bands or with very ugly things in
larger bands combined with their description is a subtle storyline of
"creation killing creator".