New Vaults monster: Icegrip Stalwart
Icegrip Stalwarts are spellcasting monsters with a new Creeping Frost
spell, which deals cold damage to and flash-freezes all enemies adjacent
to walls. They also have an AF_COLD touch attack. Overall, they're
intended to be a mid-tier threat which plays with the terrain of Vaults
(where there tend to be a lot of walls around!) and which fills a gap in
rC-checking monsters in Vaults.