Nerf the dex meta
Feedback has made it clear that it's a bit too easy to reach high levels
of ev.  This commit reduces the dodging*dex factor by 20%. High levels
of dex and dodging should remain rewarding, but a little less so, and
will be worse until the player gets there. To compensate the very early
game, the flat portion of dodge_bonus increases slightly.

[ Closes #1784. Committer's note: squashed and rebased. ]