Don't waste a turn when there isn't enough space to conjure foxfire
All (pseudo-)summoning spells can waste a turn in this way, but
foxfire was a particularly noticable case of this because of the
the 1-tile spawn radius and the short duration/self-destruction,
allowing monsters to block all potential spawn locations more
easily.

This implementation still wastes a turn if the player miscasts
the spell with no potential locations for a foxfire to spawn. A
better solution would be to create a helper function for all
summoning spells that pre-checks all potential spawn locations,
aborting the spell if none are available, but this is beyond the
scope of this commit.