Nerf new Ashenzari skill boost
Making it easier to get multiple of the same curse and also doubling the
boost formula was effective in making it more powerful. Too powerful.

Two curses of the same type still give a multiplier equal to the "fully
bound" level of old Ashenzari, so hopefully this lands on a good power
level.