Reduce explicit species checks: draconian
This commit is part of a series that attempts to more cleanly treat
species as bundles of traits/behaviors that have their own well-defined
semantics, and minimize the use of explicit species checks in the code.

Red draconians: fix up a manual zap conversion.

Pale draconians: convert steam resistance to a real mut (before it
wasn't shown at all, perhaps because it seems not-very-important?)

Grey draconians:
 * There's still a hard-coded ac bonus in player::racial_ac, wasn't
   obvious to me that there is a clean way to abstract this one.
 * Their relationship to water was very confused; I have (maybe) cleaned
   this up a bit. Previously, they had MUT_UNBREATHING, a fakemut about
   walking through water, can_swim=True, and a hard-coded check in
   player::extra_balanced. The latter, I'm pretty sure, is currently
   redundant with swimming (it probably wasn't in the past). I have
   consolidated these by making MUT_UNBREATHING two levels, where the
   second level is flavored like the old fakemut. The fakemut is gone,
   they aren't specified as swimming (though the mut is still checked in
   swimming code), and extra_balanced has been refactored, including
   some unrelated cleanups.