Don't make a note when the player finds an altar for a god the character can't worship.
Add a "temp" parameter to player_can_join_god() and player::undead_or_demonic(). If false, the functions ignore temporary effects from mutations and spells, and only consider permanent barriers such as the player's species and Ru sacrifices.

Use this to stop the game from making a note on finding an altar for a god the character will never be able to worship.