Reduce explicit species checks: gnoll
All that was needed for this is to convert distributed training from a
fakemut to a real mutation. This was fairly straightforward overall.
`_player_is_gnoll` from e00f04a389b562 becomes a direct mutation check;
it looks like this helper function was added to make the code overall
more readable/systematic (probably to avoid bugs when modifying this
code) so I think checking a mutation in a systematic way accomplishes
this goal as well.