Abort foxfire if it would obviously have no effect (AdamPG)
Check if there's a visibly open space and abort the spell if there isn't
one. This check happens before spell failure; if foxfire fizzles after
this check it reveals an invisible foe, which the player should be
charged for.

Detected invisible foes currently don't count which is an edge case in a
variety of ways in crawl, but a fix out of scope for this small tweak.

Closes #1790