Clean up miasma immunity
* Remove the redundant and now inert (after 7c3aafa790)
  MUT_MIASMA_IMMUNITY from gargoyles and from the game.
* Simplify the way unbreathing interacts with rMiasma, by just moving it
  straight into that code. (This doesn't affect gargoyles, because they
  are also non-living, but may affect grey draconians.)

The fact that gargoyles are non-living isn't really cued anywhere, but
the unbreathing part is shown in `A`, and since this now absolutely
entails rMiasma also, it's probably ok to not directly show that effect
in `A`. The wording of non-breathing could be tweaked if its cloud
immunities aren't clear enough. (It's possible there's not even a need
for non-living as such to prevent miasma now?)