Clean up rTorment muts a bit
This converts MUT_TORMENT_RESISTANCE to a two-level mutation. Level 1 is
50% damage reduction, previously hard-coded to gargoyles. The Gr fakemut
becomes this. Level 2 is previous level 1, and so mummies/ghouls move to
level 2.

MUT_STOCHASTIC_TORMENT_RESISTANCE independently exists as a somewhat
weird and unusual resistance mechanism that on balance seems to me to be
a bit worse than just simple level 1 as it now exists, though it allows
for getting lucky. I'll leave it alone for now though.