Adjust ghost vault allow_dup tagging
Previously we marked all ghost vaults as allow_dup, which meant they
could all place multiple times. This was mostly a relic of when there
were fewer ghost vaults and we didn't want games to be stuck without a
ghost vault available. Now we have a wide variety of ghost vaults,
including a reasonable set with both wide placement and generic enough
theme that we can designate with allow_dup to prevent ghost vault
shortages.

This commit removes allow_dup as a default ghost vault tag, explicitly
giving it to a small number of ghost vaults instead.