Remove the holy word effect from Sanctuary
Zin's Sanctuary did two things: 1. blast holy word and 2. create a
sanctuary with strict non-agression conduct (even for evil things).
Which mean that beyond flavour, the player couldn't take advantage of 1,
and in any case this was never mentioned in the description in-game.

Sanctuary is a defensive ability, so instead of also doing damage (that
if non-lethal can't be followed up on sensibly) just create the
sanctuary.

As an added benefit, Closes #1703