Update Elf Vault
New changes! We have a map of their camp (dead in crypt).
When the elves die, there's a chance of a corpse eater (or a spell caster in crypt).
Made Depths elves more dangerous but gave loot as compensation.
Elves now have walking sticks so they don't get as tired.