Convert octopode tentacles to a real mutation
These were a mess of special cases and custom form checks, so it's
cleaner to have these as a real mutation. There's a minor loss of info
in the % screen as "constrict 8" is replaced by just "tentacles", but
this custom status could be readded if it's too missed.

This leaves only the Op armour gimmick (restrictions + mantle) as
hard-coded for this species.