Swap L/R ring in % display order
efe1b44 accidentally changed which finger the first ring goes on. This
is actually good, because it resolves 11879 -- the first ring and the
sacrifice hand are drawn from a consistent order where they are always
different. However, now the order in % with only one ring looks odd.
This prints the right ring first. (Sorry lefties..)