Refactor vampirism as a mutation
This commit is part of a series that attempts to more cleanly treat
species as bundles of traits/behaviors that have their own well-defined
semantics, and minimize the use of explicit species checks in the code.

This commit refactors vampirism into a two-level mutation; level 1 gives
the base set of vampiric abilities that were previously conditioned on
species checks, and level 2 gives batform as well. I wasn't sure if this
would be worth the effort, but I think in the end it's a much cleaner
design than before.

There's a few small related cleanups here, including to mutation display
and to item annotations. One thing of note is that I marked torment as a
good_item when you are torment immune, which may be initially confusing
to people, because the only thing that previously got this color is
acquirement and they don't exactly form a natural class. (Emergency
items are colored yellow, which preempts good_item.)