Refactor mutation listing
Previously this was handled in two separate but parallel places: a
function in output.cc used for the `%` screen, and a more developed
version in mutation.cc that produces the long list for `A` (with a more
interpretable order that doesn't just go by the mutation enum). This
commit unifies the two by merging in the short-form listing to the
existing code in mutation.cc, leading to more consistency across the two
displays and less likelihood of them getting out of sync. The order of
display for `A` is now used for `%`.

I have further tweaked the order to consistently show temporary effects
first, and made a few other small changes mainly to terse mutation
description special cases.