Some tweaks to a vault
Adjust the threats and "elven" equipment to more appropriately scale
with depth of placement. Check for depth in Dungeon, upgrading things
for post D:13. Remove the conditional "the elves died" theme that
previously happen in any branch. I'm afraid the theme won't be well
communicated with just random monsters and some elven skeletons around.
The placement in Crypt covers a similar theme, with the elves becoming
spectral, and the vault poses a substantial threat in that branch. In
the case of Crypt, also add the elven gear as loot underneath the
respective spectral elves and fix up the dead trees.

Also add transparent, no_trap_gen, and no_monster_gen tags to this
vault.

Closes #1811.