Show form info on the `A` screen
A lot of information about forms is the same kind of thing that would
already be shown on the `A` screen (I could even imagine reimplementing
a big chunk of this as actual mutations, there's already a fair bit of
overlap). Some form-based / non-permanent (but non-tempmut) info was
already shown here, including Mf/Vp form info, beastly appendage, and a
few random religious properties. This commit generalizes this to always
show key pieces of info for forms. I haven't tried to be exhaustive
here; many things that are already shown in the `%` screen somehow don't
get their own entry.

This is implemented basically by extending the fakemut concept to forms;
there's a mix of programmatically-generated fakemuts and custom ones. As
noted above, for at least some of these properties it might be worth
considering how to generalize the beastly appendage approach.