Reduce Ash skill boost
This was still a bit overtuned. Getting to 2 curses in a category got to
the same skill boost as "fully bound" under old Ash; early on this could
already get ahead of the power curve. In the long run, since the number
of curses is uncapped, a player could potentially reach double the boost
of old Ash in a particular curse category.

Taking the time to craft focused curses is a strategic tradeoff, and
players are enjoying it so I don't want to put a hard cap on the boost.