Allow shop placement on D:1-3
Let the player see some exciting shop items earlier. This is not as much of a
power buff as it seems, since the player won't have enough gold to buy things
in most cases (though I could see banning consumable shops or raising prices,
perhaps). Shop rate per floor of D is nerfed so that the total number of
dungeon shops is slightly lower overall.