Legalize early rare scrolls.
This removes the depth_mod check for scrolls of acquirement, summoning, vuln,
silence, brand weapon, torment, and holy word. Scroll weights for those scrolls
were decreased slightly (except for vuln, which was increased) to compensate
and to make the total scroll weight nicer. I could see dropping weight further
if needed, and there is some concern about startscumming with ?acq.
Coincidentally, this despoilers scroll id; there is no longer a reason to track
the floor you found a scroll on.