Make some monster item functions more available
The function ghost_monster_weapon() had been previously used only for
ghost vaults, but this functionality is useful for any vault that wants
to specify thematic monster weapons. This commit moves the function and
associated table to dgn.monster_weapon() and fix up documentation of
both the table and the function.

Also rename random_item_def() to dgn.random_item_def(), to be consistent
with use of the dgn namespace. Adapt all the vaults currently using
these functions to the new names.