Rework an altar vault
The vault due_chaos_kobolds placed kobolds with distortion or chaos
weapons in late branches for flavor, but we're moving away from chaff
monsters in late game areas. This commit reworks the vault to be a runed
door vault that places in most of Dungeon, using depth-appropriate
Kobolds with ugly thing monsters mixed in for later depths.