Respawn of ancestor faster.
It's tiring to take care of ancestor so that they don't die. Therefore, it will be respawn faster to reduce the management burden of the player. Although respawn has become very fast, the absence of an ancestor for 5-15 turns in urgent combat situations is still threat to the player. Therefore, I expect that the value of 'idealise' will not be completely tarnished.