Update to Blade Hands description
The old Blade Hands description used the word "hands" regardless of
the manipulator type for the species. In addition, it failed to mention
that it would meld shields and gloves. A function was added to l-you.cc
to get the hand type of the player species, and the Blade Hands
description now checks hand type to determine the noun used for the
player's manipulators. The description was also changed to say that
shields and gloves are also melded.