New Monster: salamander tyrant (i-x)
Salamander tyrants are meant to complement nagaraja as a top-tier threat
in Snake, so that the latter don't have to do so much heavy lifting in
end vaults etc. They're also meant to help depths lava salamanders have
a bit more relevance.

Snake has lots of melee bruisers, so the tyrant is a damaging
spellcaster: flimsier but with some hard hitting spells it likes to cast
a lot. The Tyrant is also fine with sitting back behind tougher Naga
cousins: its spells don't require line of fire, only line of sight.

Two spells play with terrain:

Eruption fires the magma-themed upheaval flavour, hitting the player
with partially-resistible BOM flavoured damage and creating temporary
lava (constraining the player's movement).

Pyroclastic Surge strikes foes of a tyrant that are above or adjacent to
lava with pure fire damage.

The third is a debuff weakening gaze, inflicting DUR_WEAK on the player.
As a gaze attack it counts as instant and does not require line of fire,
only line of sight.

Combined with the terrain constraining attacks the gaze should make
tyrants tough to fight in melee, but in a different way than the tankier
Naga.