Updates to Hell entry and vestibule vaults
In line with the removal of weaker Depths monsters, update the Hell
entry vaults to place nothing worse than a tier 3 demon-wise and to
update other monsters where appropriate.

Likewise update the vestibule subvaults in a similar fashion, so that
players coming from Depths won't be confronted with e.g. white imps. The
Hells still place plenty of chaff monsters, but those at least have Hell
effects and other much stronger monsters we don't use the in the
vestibule.