Save abyss on AK start
Before this commit, this abyss was in a precarious state: generated, but
not saved to disk. An AK start crash (of the kind that doesn't save the
game) could load to a load that put the character in a featureless void,
crashing immediately (except in wizmode). A previous commit does fixup
for this scenario, but this commit should prevent it from happening for
future games altogether.