New "Demonic Will" mutation for the damnation facet.
This mutation provides Will+ and additionally deals a small amount of
retaliatory damage whenever the player beats a monster's will check. This
replaces rF+ as the first rank of the hurl damnation facet; the facet is now
will/res torment/hurl damnation. The theme is something like becoming a
brimstone fiend.