Refactor some arm order stuff, again
This changes how arm order is handled from efe1b4415. There, I tried to
make things more consistent by providing the arm to be sacrificed as the
last element of a vector. This unfortunately has run up against
inconsistencies with display order, where it is better to show the left
arm first (when there are two). Instead, localize the missing hand
calculation to species.cc. It might be better to change the sac
hand target, but I don't want to deal with the save compat involved.

This commit is a regression for 11879; the default ring target is now
once again the left hand (also the sac target).